hotfix for conda run -n bug
As of conda 0.48.1 `conda run -n` doesn't exit with the code that the
command it ran exited with. This causes all the Numba integration tests
to silently be successful. A PR is in flight to fix this already, but we
hotfix it for now to get the tests up and running again.

Issue:

https://github.com/conda/conda/issues/9599

PR:

https://github.com/conda/conda/pull/9665