implement git_latest_tag
This is a utility function that can be used in targets to determine the
latest version of the software from the available git tags. It uses the
`parse` method from `packaging.version` to determine the order of the
versions. If the tags are somewhat sane this does a good job. The
keyword argument `vprefix` can be used to prefix the resulting version
with a `v`, which is a common versioning scenario. The keyword argument
`exclude_filter` can be used to exclude tags, or sets  of tags that
should not be considered, for example non-parsable tags like `docs` or
`website`. Supply a callable that takes a single argument and  returns
`True` if that argument should be exluded.