pin to python 3.7 for execution
For `conda run` to work, the following patch is needed:

https://github.com/conda/conda/pull/9665

This means, the `texasbbq` package is stuck on conda=4.7. The latest
miniconda, offers Python 3.8 by default, but conda 4.7 isn't compatible
with Python 3.8 resulting in the following error during installation:

```
UnsatisfiableError: The following specifications were found
to be incompatible with the existing python installation in your environment:

Specifications:

  - conda=4.7 -> python[version='>=2.7,<2.8.0a0|>=3.6,<3.7.0a0|>=3.7,<3.8.0a0']

Your python: python=3.8
```

As a temporary workaround while we wait for a conda version that
includes the above patch, we pin Python  to 3.7.