execute conda run with --no-capture-output
The default behaviour of `conda run` is to buffer the output and then
display the entire running output after the command has completed. For
long running compile or testing processes this is suboptimal from a
usability perspective since the progress can not be monitored in
real-time. Adding the flag `--no-capture-output` means that `conda run`
will not buffer the output but stream it directly to the terminal which
improves usability.